refactor(native): move runtime tick orchestration into subfolder
